Move batch_send to common mixin
